Android: replace ndk.dir with android.ndkVersion in build.gradle
The use of ndk.dir in local.properties is deprecated in favor of
android.ndkVersion in build.gradle, and will be totatlly removed in
the future, so we need to adapt to that.

This 5.15 cherry-pick retrieve the ndk version without the use
of QSettings because we use androiddeployqt uses bootstrap
when building as a host_build while doing a cross-compilation.

Fixes: QTBUG-91391
Change-Id: I54c57113a759d43c3685c9cdf2b9dcc5c948c0fd
Reviewed-by: BogDan Vatra <bogdan@kdab.com>
Reviewed-by: Rami Potinkara <rami.potinkara@qt.io>
(cherry picked from commit 7235eeab19e7e50ba102938993acc04d14b58164)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>